rsx: Avoid calling any blocking callbacks from threads that are not rsx::thread - Defers on_notity_memory_unmapped to only run from within rsx context - Avoids passive_lock + writer_lock deadlock

+++ rpcs3/Emu/RSX/rsx_utils.h
@@ -159,7 +159,7 @@ namespace rsx
 					{
 						// Head
 						write_offset = (offset - base_offset);
-						write_length = std::min<u32>(block.second - write_offset, remaining_bytes);
+						write_length = std::min<u32>(block.second - write_offset, (u32)remaining_bytes);
 					}
 					else if (base_offset > offset && block_end <= write_end)
 					{
@@ -171,7 +171,7 @@ namespace rsx
 					{
 						// Tail
 						write_offset = 0;
-						write_length = remaining_bytes;
+						write_length = (u32)remaining_bytes;
 					}
 					else
 					{
